Created function that opens links at desired time based on form input
Currently, the extension can consume a link and date+time, and an alarm will be created to be fired at that time. When fired, the alarm opens the link in a new tab. Next steps are to store alarms so that they remain for future sessions, and present all alarms on the popup interface.