Added remove button, created visual representation of link alarms
Used javascript to create a list that presents all links with their scheduled times. Also added a button to each item of the list to remove scheduled links.  The list updates when a link is added, removed, when all alarms are cleared and when the extension reopens. Created a function that adds a hash mark after a URL along with a number to differentiate from identical URLs to bypass the limitation of alarms having unique names/links in chrome.alarms API.